fix(case): rename model module to support_case to avoid SQL case() collision
